# PERF: nancorr_spearman fastpath