@jreback 2021-06-09
can you  add this PR number to the existing perf note on corr, ping on green.