@mzeitlin11 2021-06-09
Added a small fix and test in latest commit for edge case of min periods > frame length